requirements.txt
claude-agent-sdk>=0.1.60
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
apscheduler>=3.10.0
yfinance>=0.2.40
pandas>=2.0.0
pandas-market-calendars>=4.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jinja2>=3.1.0
python-multipart>=0.0.9
httpx>=0.27.0
requests>=2.31.0
python-dotenv>=1.0.0
aiosmtplib>=3.0.0
python-telegram-bot>=20.7
pytz>=2024.1
